FEAT : 추가중인 TestCaseInput도 삭제 가능하도록 기능 추가
