+++ src/components/solve/ProblemDescription/common.tsx
@@ -1,5 +1,6 @@
 import { styled } from "styled-components";
 import { spacer, SpacerProps } from "../../../lib/spacer";
+import { ReactNode, useState } from "react";
 
 export interface TestCase {
   input: string;
@@ -52,5 +53,84 @@ export const TableItem = styled(TableHeader)`
   }
   > td:focus-within {
     background-color: #e6e6e6;
+  }
+`;
+
+interface DeletableTableItemProps {
+  index: number;
+  deleteItem: () => void;
+  children: ReactNode;
+}
+
+export function DeletableTableItem({
+  index,
+  deleteItem,
+  children,
+}: DeletableTableItemProps) {
+  const [isHover, setIsHover] = useState<boolean>(false);
+
+  return (
+    <DeletableTableRow
+      onMouseEnter={() => {
+        setIsHover(true);
+      }}
+      onMouseLeave={() => {
+        setIsHover(false);
+      }}
+    >
+      <DeletableTh>
+        {isHover ? (
+          <button onClick={deleteItem}>삭제</button>
+        ) : (
+          <BoldText>{index}</BoldText>
+        )}
+      </DeletableTh>
+      {children}
+    </DeletableTableRow>
+  );
+}
+
+const DeletableTableRow = styled(TableHeader)`
+  td {
+    background-color: #f6f6f6;
+    padding: 10px 15px;
+    border-radius: 5px;
+  }
+`;
+
+const DeletableTh = styled.th`
+  background-color: transparent;
+  > button,
+  p {
+    width: 100%;
+    height: 100%;
+    color: #323232;
+    border-radius: 5px;
+    animation: appear 300ms;
+    @keyframes appear {
+      from {
+        opacity: 0;
+      }
+    }
+  }
+  > button {
+    display: block;
+    padding: 13.5px 17.5px;
+    font-size: 14px;
+    font-weight: bold;
+    line-height: 160%;
+    letter-spacing: 5%;
+    border: none;
+    background-color: #f0745f;
+    transition: 0.3s;
+    &:hover {
+      transition: inherit;
+      background-color: #e6e6e6;
+      cursor: pointer;
+    }
+  }
+  > p {
+    padding: 10px 15px;
+    background-color: #f6f6f6;
   }
 `;
